split up the original single-file addon into a multi-file addon for better maintainability
